实体管理

+++ src/main/java/com/zlhhh/networksecurity/service/impl/EntityInfoServiceImpl.java
@@ -12,6 +12,7 @@ import javax.annotation.Resource;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 
 @Service
 public class EntityInfoServiceImpl extends ServiceImpl<EntityInfoMapper, EntityInfo> implements EntityInfoService {
@@ -43,7 +44,7 @@ public class EntityInfoServiceImpl extends ServiceImpl<EntityInfoMapper, EntityI
      */
     @Override
     public String createNode(Integer entityId) {
-        EntityInfo entityInfo = queryEntity(entityId, false);
+        EntityInfo entityInfo = queryEntity(entityId);
         String label = entityInfo.getLabel();
         Map<String, Object> entityInfoMap = generateHashMap(entityInfo);
         System.out.println(entityInfoMap);
@@ -52,13 +53,14 @@ public class EntityInfoServiceImpl extends ServiceImpl<EntityInfoMapper, EntityI
 
     /**
      * 更改节点的值
+     * 这个方法不用了
      *
      * @param entityId 节点id(mysql)
      * @return
      */
     @Override
     public String setNode(Integer entityId) {
-        EntityInfo entityInfo = queryEntity(entityId, false);
+        EntityInfo entityInfo = queryEntity(entityId);
         String label = entityInfo.getLabel();
         String value = entityInfo.getValue();
         Map<String, Object> entityInfoMap = generateHashMap(entityInfo);
@@ -69,18 +71,93 @@ public class EntityInfoServiceImpl extends ServiceImpl<EntityInfoMapper, EntityI
     }
 
     /**
+     * 从mysql中根据id取到实体数据
+     * 进行判断：
+     * 如果neo4j_id是空
+     * 则新建节点
+     * 否则更新节点数据doSetNodeInfo
+     *
+     * @param entityIds idList
+     * @return String
+     */
+    @Override
+    public String setNodeInfo(List<Integer> entityIds) {
+        String entityValues = "";
+        for (Integer entityId : entityIds) {
+            EntityInfo entityInfo = queryEntity(entityId);
+            String entityValue;
+            if (entityInfo.getNeo4jId() == null) {
+                entityValue = doCreateNode(entityInfo);
+            } else {
+                entityValue = doSetNodeInfo(entityInfo);
+            }
+            entityValues.concat(", ");
+            entityValues.concat(entityValue);
+        }
+        return entityValues;
+    }
+
+    /**
      * 删除节点
      *
      * @param entityId 实体在mysql中的id
      */
     @Override
     public void deleteNode(Integer entityId) {
-        EntityInfo entityInfo = queryEntity(entityId, true);
+        EntityInfo entityInfo = queryEntity(entityId);
         String label = entityInfo.getLabel();
+        if (entityInfo.getNeo4jId() != null) {  // 证明这个实体在neo4j中有对应的节点，不是新增的，需要更新neo4j中的信息
+            Long neo4jId = entityInfo.getNeo4jId();
+            neo4jEntityRepository.deleteNodeById(label, neo4jId);
+        }
+        entityInfo.setIsDeleted(true);
+        entityInfoMapper.updateById(entityInfo);
+//        Long neo4jId = neo4jEntityRepository.searchNodeIdByNameAndLabel(label, value);
+//        System.out.println(neo4jId);
+    }
+
+    @Override
+    public void deleteNodeBatch(List<Integer> entityIds) {
+        for (Integer entityId : entityIds) {
+            deleteNode(entityId);
+        }
+    }
+
+    /***************/
+
+    private String doCreateNode(EntityInfo entityInfo) {
+        String label = entityInfo.getLabel();
+        Map<String, Object> entityInfoMap = generateHashMap(entityInfo);
+        String entityValue = neo4jEntityRepository.createNode(label, entityInfoMap);
+        Long neo4jId = neo4jEntityRepository.searchNodeIdByNameAndLabel(label, entityValue);
+        entityInfo.setNeo4jId(neo4jId);
+        entityInfo.setUpdated(true);
+        entityInfoMapper.updateById(entityInfo);
+        return entityValue;
+    }
+
+    private String doSetNodeInfo(EntityInfo entityInfo) {
+        String label = entityInfo.getLabel();
+        Long id = entityInfo.getNeo4jId();
         String value = entityInfo.getValue();
-        Long neo4jId = neo4jEntityRepository.searchNodeIdByNameAndLabel(label, value);
-        System.out.println(neo4jId);
-        neo4jEntityRepository.deleteNodeById(label, neo4jId);
+        if (Objects.equals(label, "Organization")) {
+            String introduction = entityInfo.getIntroduction();
+            String occurtime = entityInfo.getOccurtime();
+            String referlink = entityInfo.getReferlink();
+            String motivation = entityInfo.getMotivation();
+            entityInfo.setUpdated(true);
+//            System.out.println(entityInfo);
+            entityInfoMapper.updateById(entityInfo);
+            return neo4jEntityRepository.setOrganizationNode(label, id, value, introduction, occurtime, motivation, referlink);
+        } else if (Objects.equals(label, "Attacktype")) {
+            String introduction = entityInfo.getIntroduction();
+            return neo4jEntityRepository.setAttackTypeNode(label, id, value, introduction);
+        } else if (Objects.equals(label, "Sha256")) {
+            String format = entityInfo.getFormat();
+            return neo4jEntityRepository.setSha256Node(label, id, value, format);
+        } else {
+            return neo4jEntityRepository.setOtherNode(label, id, value);
+        }
     }
 
     /**
@@ -118,7 +195,7 @@ public class EntityInfoServiceImpl extends ServiceImpl<EntityInfoMapper, EntityI
      * @param entityId
      * @return
      */
-    private EntityInfo queryEntity(Integer entityId, boolean isDelete) {
+    private EntityInfo queryEntity(Integer entityId) {
         EntityInfo entityInfo = entityInfoMapper.selectById(entityId);
 //        LambdaQueryWrapper<EntityInfo> lambdaQueryWrapper = new LambdaQueryWrapper<>();
 //        lambdaQueryWrapper.eq(EntityInfo::getId, entityId);
@@ -132,10 +209,8 @@ public class EntityInfoServiceImpl extends ServiceImpl<EntityInfoMapper, EntityI
 //            throw new ServiceException(Constants.CODE_500, "系统错误");
 //        }
         entityInfo.setUpdated(true);
-        if (isDelete) {
-            entityInfo.setIsDeleted(true);
-            entityInfoMapper.updateById(entityInfo);
-        }
+        entityInfoMapper.updateById(entityInfo);
+
         return entityInfo;
     }
 }
